Fixed a bug with GPUDevice::MapBuffer()

+++ Graphics/CommandBufferRing.h
@@ -31,12 +31,12 @@ public:
     CommandBuffer* GetCommandBuffer(uint32 frame, bool begin);
     CommandBuffer* GetCommandBufferInstant(uint32 frame, bool begin);
 
-    static uint32 PoolFromIndex(uint32 index) { return index / BUFFER_PER_POOL; }
+    static uint16 PoolFromIndex(uint32 index) { return (uint16)index / BUFFER_PER_POOL; }
 
-    static const uint32 MAX_THREADS = 1;
-    static const uint32 MAX_POOLS = MAX_SWAPCHAIN_IMAGES * MAX_THREADS;
-    static const uint32 BUFFER_PER_POOL = 4;
-    static const uint32 MAX_BUFFERS = BUFFER_PER_POOL * MAX_POOLS;
+    static const uint16 MAX_THREADS = 1;
+    static const uint16 MAX_POOLS = MAX_SWAPCHAIN_IMAGES * MAX_THREADS;
+    static const uint16 BUFFER_PER_POOL = 4;
+    static const uint16 MAX_BUFFERS = BUFFER_PER_POOL * MAX_POOLS;
 
     GPUDevice* gpu_device;
     VkCommandPool vk_command_pools[MAX_POOLS];
